QR_code_scaning/requirements.txt
qrcode==7.3.1
opencv-python==4.5.5.64
pillow==9.1.1
requests==2.26.0
django==3.2.9
pyopenssl==21.0.0